+++ server/webpack.config.js
@@ -13,5 +13,19 @@ module.exports = withDefaults({
     output: {
         filename: 'esphomeServer.js',
         path: path.join(__dirname, 'out')
+    },
+    externals: {
+        'prettier': 'prettier',
+        'bufferutil': 'bufferutil',
+        'utf-8-validate': 'utf-8-validate',
+        'yaml-language-server': 'yaml-language-server/lib/umd',
+        'yaml-language-server-parser': 'yaml-language-server-parser',
+        'vscode-json-languageservice': 'vscode-json-languageservice',
+        'jsonc-parser': 'jsonc-parser',
+        'vscode-uri': 'vscode-uri',
+        'vscode-languageserver-textdocument': 'vscode-languageserver-textdocument',
+        'vscode-languageserver-types': 'vscode-languageserver-types',
+        'js-yaml': 'js-yaml',
+        'vscode-nls': 'vscode-nls',
     }
 });